requirements.txt
asgiref==3.8.1
boto3==1.38.18
botocore==1.38.18
coverage==7.6.12
dj-database-url==2.3.0
django==5.2.1
django-bootstrap5==25.1
django-extensions==3.2.3
django-s3-storage==0.15.0
freezegun==1.5.1
gunicorn==23.0.0
jmespath==1.0.1
model-bakery==1.20.4
packaging==25.0
pillow==11.1.0
psycopg==3.2.5
psycopg-binary==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ruff==0.11.10
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
